feat: implementado genericos e modais (falta checar ainda)
