+++ src/finance/components/form/FinanceForm.css
@@ -1,0 +1,160 @@
+.finance-form {
+  display: grid;
+  gap: 1rem;
+}
+
+.finance-form__grid {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 1rem;
+}
+
+.finance-field {
+  display: grid;
+  gap: 0.4rem;
+}
+
+.finance-field.is-wide {
+  grid-column: 1 / -1;
+}
+
+.finance-field__label {
+  color: var(--finance-muted);
+  font-size: 0.78rem;
+  font-weight: 800;
+  letter-spacing: 0.03em;
+}
+
+.finance-field__hint {
+  display: block;
+  margin: -0.05rem 0 0.1rem;
+  color: var(--finance-muted);
+  font-size: 0.76rem;
+  font-weight: 700;
+  line-height: 1.4;
+}
+
+.finance-field__hint--danger {
+  color: var(--finance-danger);
+}
+
+.finance-field input,
+.finance-field select,
+.finance-select select {
+  width: 100%;
+  min-height: 2.8rem;
+  padding: 0 0.8rem;
+  border: 1px solid var(--finance-border-strong);
+  border-radius: 10px;
+  outline: none;
+  background: var(--finance-surface-solid);
+  color: var(--finance-text);
+  font: inherit;
+}
+
+.finance-field input:focus,
+.finance-field select:focus,
+.finance-select select:focus {
+  border-color: var(--finance-primary);
+  box-shadow: 0 0 0 3px color-mix(in srgb, var(--finance-focus) 14%, transparent);
+}
+
+.finance-select {
+  position: relative;
+}
+
+.finance-select select {
+  appearance: none;
+  padding-right: 2rem;
+}
+
+.finance-select__chevron {
+  position: absolute;
+  top: 50%;
+  right: 0.7rem;
+  color: var(--finance-muted);
+  font-size: 0.72rem;
+  transform: translateY(-50%);
+  pointer-events: none;
+}
+
+.finance-checkbox-row {
+  display: flex;
+  align-items: center;
+  gap: 0.65rem;
+  color: var(--finance-text);
+  font-size: 0.92rem;
+  font-weight: 600;
+}
+
+.finance-checkbox-row input {
+  width: 1.05rem;
+  height: 1.05rem;
+  accent-color: var(--finance-primary);
+}
+
+.finance-modal-actions {
+  display: flex;
+  justify-content: flex-end;
+  gap: 0.7rem;
+  margin-top: 0.4rem;
+}
+
+.finance-button {
+  min-height: 2.75rem;
+  padding: 0.65rem 1rem;
+  border: 1px solid transparent;
+  border-radius: 10px;
+  font: inherit;
+  font-weight: 800;
+  cursor: pointer;
+}
+
+.finance-button--primary {
+  background: var(--finance-primary);
+  color: var(--finance-on-strong);
+}
+
+.finance-button--primary:hover,
+.finance-button--primary:focus-visible {
+  background: var(--finance-primary-strong);
+}
+
+.finance-button--secondary {
+  border-color: var(--finance-border-strong);
+  background: var(--finance-surface-solid);
+  color: var(--finance-text);
+}
+
+.finance-button--secondary:hover,
+.finance-button--secondary:focus-visible {
+  background: var(--finance-hover);
+}
+
+.finance-button--danger {
+  background: var(--finance-danger);
+  color: var(--finance-on-strong);
+}
+
+.finance-button--danger:hover,
+.finance-button--danger:focus-visible {
+  background: var(--finance-danger-strong);
+}
+
+@media (max-width: 720px) {
+  .finance-form__grid {
+    grid-template-columns: 1fr;
+  }
+
+  .finance-field.is-wide {
+    grid-column: auto;
+  }
+
+  .finance-modal-actions {
+    flex-direction: column-reverse;
+  }
+
+  .finance-modal-actions .finance-button {
+    width: 100%;
+  }
+}
