+++ src/finance/components/modal/FinanceModal.css
@@ -1,0 +1,67 @@
+.finance-modal__overlay {
+  position: fixed;
+  z-index: 1000;
+  inset: 0;
+  display: grid;
+  place-items: center;
+  padding: 1rem;
+  background: color-mix(in srgb, var(--color-text-primary) 42%, transparent);
+  backdrop-filter: blur(2px);
+}
+
+.finance-modal {
+  width: min(100%, 680px);
+  max-height: min(90vh, 820px);
+  overflow: auto;
+  border-radius: 22px;
+  background: var(--finance-surface-solid);
+  color: var(--finance-text);
+  box-shadow: var(--finance-shadow-elevated);
+}
+
+.finance-modal--small {
+  width: min(100%, 500px);
+}
+
+.finance-modal--large {
+  width: min(100%, 860px);
+}
+
+.finance-modal__header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+  margin: 0 1.2rem;
+  padding: 1.2rem 0 1rem;
+  border-bottom: 1px solid var(--finance-border);
+}
+
+.finance-modal__header h2 {
+  margin: 0;
+  font-size: 1.25rem;
+}
+
+.finance-modal__close {
+  display: grid;
+  width: 2.45rem;
+  height: 2.45rem;
+  flex: 0 0 2.45rem;
+  place-items: center;
+  border: 0;
+  border-radius: 50%;
+  background: transparent;
+  color: var(--finance-muted);
+  font: inherit;
+  cursor: pointer;
+}
+
+.finance-modal__close:hover,
+.finance-modal__close:focus-visible {
+  background: var(--finance-hover);
+  color: var(--finance-primary);
+}
+
+.finance-modal__body {
+  padding: 1.1rem 1.2rem 1.2rem;
+}
