+++ src/finance/components/page/FinancePage.css
@@ -1,0 +1,128 @@
+.finance-page,
+.finance-surface-theme {
+  --finance-bg: var(--list-page-bg, var(--color-bg));
+  --finance-surface: var(--list-shell-surface, var(--color-surface));
+  --finance-surface-solid: var(--color-surface);
+  --finance-soft: var(--list-card-muted-surface, var(--color-surface-muted));
+  --finance-filter: var(--color-surface-muted);
+  --finance-border: var(--list-border-soft, var(--color-border));
+  --finance-border-strong: var(--color-border);
+  --finance-text: var(--color-text-primary);
+  --finance-muted: var(--color-text-secondary);
+  --finance-primary: var(--color-primary);
+  --finance-primary-strong: var(--color-primary-hover);
+  --finance-danger: var(--error-color);
+  --finance-danger-strong: var(--error-color-hover);
+  --finance-success: var(--success-color);
+  --finance-focus: var(--color-focus);
+  --finance-hover: var(--list-hover-surface, var(--color-surface-muted));
+  --finance-danger-bg: var(--error-bg);
+  --finance-danger-border: var(--error-border-color);
+  --finance-shadow: var(--shadow-card);
+  --finance-shadow-elevated: var(--shadow-elevated);
+  --finance-on-strong: var(--color-surface);
+}
+
+[data-theme="dark"] .finance-page,
+[data-theme="dark"] .finance-surface-theme {
+  --finance-on-strong: var(--color-text-primary);
+}
+
+body:has(.finance-page),
+#root:has(.finance-page) {
+  background: var(--list-page-bg, var(--color-bg));
+}
+
+.finance-page {
+  --finance-header-offset: var(--app-header-height, var(--header-height, 5.5rem));
+  position: relative;
+  width: 100%;
+  min-height: 0;
+  height: auto;
+  padding: calc(var(--finance-header-offset) + 1rem) clamp(0.75rem, 1.5vw, 1.5rem) 1rem;
+  box-sizing: border-box;
+  overflow: visible;
+  background: transparent;
+  color: var(--finance-text);
+  font-family: Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif;
+}
+
+.finance-page,
+.finance-page * {
+  box-sizing: border-box;
+}
+
+.finance-page__panel {
+  width: 100%;
+  max-width: none;
+  margin: 0 auto;
+  overflow: visible;
+  border: 1px solid var(--finance-border);
+  border-radius: 34px;
+  background: var(--finance-surface);
+  box-shadow: var(--finance-shadow);
+}
+
+.finance-page-header {
+  padding: 1.55rem 2rem 1.35rem;
+  border-bottom: 1px solid var(--finance-border);
+}
+
+.finance-page-header > div {
+  display: grid;
+  gap: 0.25rem;
+}
+
+.finance-page-header__eyebrow {
+  color: var(--finance-muted);
+  font-size: 0.72rem;
+  font-weight: 800;
+  letter-spacing: 0.14em;
+  text-transform: uppercase;
+}
+
+.finance-page-header h1 {
+  margin: 0;
+  font-size: clamp(1.45rem, 2.2vw, 2rem);
+  line-height: 1.1;
+  letter-spacing: -0.025em;
+}
+
+.finance-page__content {
+  padding: 1rem 1rem 0.9rem;
+}
+
+[data-theme="dark"] .finance-page input[type="date"]::-webkit-calendar-picker-indicator {
+  filter: invert(1) brightness(2);
+  opacity: 1;
+}
+
+@supports (-moz-appearance: none) {
+  [data-theme="dark"] .finance-page input[type="date"] {
+    color-scheme: dark;
+  }
+}
+
+@media (max-width: 980px) and (min-width: 721px) {
+  .finance-page__content {
+    padding-inline: 0.65rem;
+  }
+}
+
+@media (max-width: 720px) {
+  .finance-page {
+    padding: calc(var(--finance-header-offset) + 0.75rem) 0.6rem 0;
+  }
+
+  .finance-page__panel {
+    border-radius: 22px;
+  }
+
+  .finance-page-header {
+    padding: 1.1rem 1rem;
+  }
+
+  .finance-page__content {
+    padding: 0.8rem;
+  }
+}
