+++ src/finance/pages/expenselist/expenseList.constants.js
@@ -1,0 +1,120 @@
+export const MONTHS = [
+  "JAN", "FEV", "MAR", "ABR", "MAI", "JUN",
+  "JUL", "AGO", "SET", "OUT", "NOV", "DEZ",
+];
+
+export const PAYMENT_TYPES = ["À vista", "Parcelado", "Recorrente"];
+
+export const PAYMENT_MODES = [
+  "Boleto",
+  "Carteira Digital",
+  "Cartão Pré-pago",
+  "Cartão de Crédito",
+  "Cartão de Débito",
+  "Cheque",
+  "Criptomoeda",
+  "Depósito Bancário",
+  "Dinheiro",
+  "Pix",
+  "Transferência Bancária",
+];
+
+export const CATEGORIES = [
+  "Infraestrutura",
+  "Mensalidade",
+  "Suprimentos",
+  "Tecnologia",
+  "Marketing",
+  "Impostos",
+  "Serviços",
+  "Pessoal",
+  "Viagens",
+  "Outros",
+];
+
+export const DEMO_EXPENSES = [
+  {
+    id: "exp-001",
+    date: "2026-05-01",
+    description: "#OS-8821",
+    payee: "Ana Costa & Associados",
+    category: "Infraestrutura",
+    value: 180,
+    paymentType: "À vista",
+    paymentMode: "Boleto",
+    paid: true,
+    attachments: [],
+  },
+  {
+    id: "exp-002",
+    date: "2026-05-03",
+    description: "#OS-8821",
+    payee: "Visa Crédito Brasil",
+    category: "Mensalidade",
+    value: 180,
+    paymentType: "Recorrente",
+    paymentMode: "Pix",
+    paid: false,
+    attachments: [],
+  },
+  {
+    id: "exp-003",
+    date: "2026-05-14",
+    description: "#OS-8823",
+    payee: "Juliana Lopes de Almeida",
+    category: "Suprimentos",
+    value: 180,
+    paymentType: "À vista",
+    paymentMode: "Cartão de Débito",
+    paid: false,
+    attachments: [],
+  },
+  {
+    id: "exp-004",
+    date: "2026-05-27",
+    description: "#OS-8821",
+    payee: "Global Soluções LTDA",
+    category: "Tecnologia",
+    value: 180,
+    paymentType: "À vista",
+    paymentMode: "Cartão de Crédito",
+    paid: true,
+    attachments: [],
+  },
+  {
+    id: "exp-005",
+    date: "2026-05-29",
+    description: "Hospedagem do site",
+    payee: "Cloud Services Brasil",
+    category: "Tecnologia",
+    value: 349.9,
+    paymentType: "Recorrente",
+    paymentMode: "Cartão de Crédito",
+    paid: true,
+    attachments: [],
+  },
+  {
+    id: "exp-006",
+    date: "2026-05-30",
+    description: "Materiais de escritório",
+    payee: "Papelaria Central",
+    category: "Suprimentos",
+    value: 226.5,
+    paymentType: "À vista",
+    paymentMode: "Pix",
+    paid: false,
+    attachments: [],
+  },
+];
+
+export const EXPENSE_TABLE_COLUMNS = [
+  { key: "date", label: "Data", width: "8.5%" },
+  { key: "description", label: "Descrição", width: "14%" },
+  { key: "payee", label: "Pago a", width: "13.5%" },
+  { key: "category", label: "Categoria", width: "11.5%" },
+  { key: "value", label: "Valor", width: "9.5%" },
+  { key: "paymentType", label: "Tipo pagamento", width: "12%" },
+  { key: "paymentMode", label: "Modo do pagamento", width: "12.5%" },
+  { key: "paid", label: "Pago?", width: "8.5%" },
+  { key: "actions", label: null, ariaLabel: "Ações", sortable: false, width: "10%" },
+];
