+++ src/finance/pages/expenselist/modals/ExpenseModalContent.css
@@ -1,0 +1,162 @@
+.expense-modal__details-card {
+  display: grid;
+  gap: 0.4rem;
+  margin-bottom: 1rem;
+  padding: 1.1rem;
+  border-radius: 16px;
+  background: var(--finance-hover);
+}
+
+.expense-modal__details-card--compact {
+  margin-bottom: 0;
+}
+
+.expense-modal__details-card span {
+  color: var(--finance-muted);
+  font-size: 0.82rem;
+  font-weight: 700;
+}
+
+.expense-modal__details-card strong {
+  color: var(--finance-primary);
+  font-size: 1.8rem;
+}
+
+.expense-modal__details-list {
+  display: grid;
+  gap: 0;
+  margin: 0;
+}
+
+.expense-modal__details-list > div {
+  display: grid;
+  grid-template-columns: 10rem 1fr;
+  gap: 1rem;
+  padding: 0.85rem 0;
+  border-bottom: 1px solid var(--finance-border);
+}
+
+.expense-modal__details-list dt {
+  color: var(--finance-muted);
+  font-weight: 700;
+}
+
+.expense-modal__details-list dd {
+  margin: 0;
+  color: var(--finance-text);
+  font-weight: 600;
+}
+
+.expense-modal__attachments {
+  display: grid;
+  gap: 1rem;
+}
+
+.expense-modal__upload-box {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 0.65rem;
+  min-height: 5.5rem;
+  padding: 1rem;
+  border: 1.5px dashed color-mix(in srgb, var(--finance-primary) 65%, var(--finance-border));
+  border-radius: 14px;
+  background: var(--finance-soft);
+  color: var(--finance-primary);
+  font-weight: 700;
+  cursor: pointer;
+}
+
+.expense-modal__upload-box input {
+  display: none;
+}
+
+.expense-modal__attachments ul {
+  display: grid;
+  gap: 0.55rem;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+}
+
+.expense-modal__attachments li {
+  display: grid;
+  grid-template-columns: auto minmax(0, 1fr) auto;
+  align-items: center;
+  gap: 0.7rem;
+  padding: 0.75rem 0.9rem;
+  border: 1px solid var(--finance-border);
+  border-radius: 12px;
+}
+
+.expense-modal__attachments li > svg {
+  color: var(--finance-primary);
+}
+
+.expense-modal__attachments li button {
+  display: grid;
+  width: 2rem;
+  height: 2rem;
+  place-items: center;
+  border: 0;
+  border-radius: 50%;
+  background: transparent;
+  color: var(--finance-danger);
+  cursor: pointer;
+}
+
+.expense-modal__attachments li button:hover,
+.expense-modal__attachments li button:focus-visible {
+  background: var(--finance-danger-bg);
+}
+
+.expense-modal__muted {
+  margin: 0;
+  color: var(--finance-muted);
+  text-align: center;
+}
+
+.expense-modal__delete-copy {
+  display: grid;
+  justify-items: center;
+  gap: 0.75rem;
+  padding: 0.4rem 0 1rem;
+  text-align: center;
+}
+
+.expense-modal__danger-icon {
+  display: grid;
+  width: 3.6rem;
+  height: 3.6rem;
+  place-items: center;
+  border-radius: 18px;
+  background: var(--finance-danger-bg);
+  color: var(--finance-danger);
+  font-size: 1.35rem;
+}
+
+.expense-modal__delete-copy p {
+  margin: 0;
+  font-size: 1rem;
+}
+
+.expense-modal__delete-copy > span {
+  max-width: 28rem;
+  color: var(--finance-muted);
+  font-size: 0.86rem;
+  line-height: 1.5;
+}
+
+.expense-modal__delete-copy code {
+  padding: 0.1rem 0.3rem;
+  border-radius: 5px;
+  background: var(--finance-soft);
+  font-size: 0.8rem;
+}
+
+@media (max-width: 720px) {
+  .expense-modal__details-list > div {
+    grid-template-columns: 1fr;
+    gap: 0.25rem;
+  }
+}
